requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
black==19.10b0
click==7.1.2
colorama==0.4.3
Django==3.0.8
django-filter==2.3.0
djangorestframework==3.11.0
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.2.2
mccabe==0.6.1
pathspec==0.8.0
pylint==2.5.3
pytz==2020.1
regex==2020.6.8
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
zipp==3.1.0